Moved all SQLUtils tests to its own project on main solution
