+++ src/OrleansProviders/Properties/AssemblyInfo.cs
@@ -23,3 +23,4 @@ using System.Runtime.InteropServices;
 [assembly: InternalsVisibleTo("Orleans.NonSiloTests")]
 [assembly: InternalsVisibleTo("TestExtensions")]
 [assembly: InternalsVisibleTo("Tester.AzureUtils")]
+[assembly: InternalsVisibleTo("Tester.SQLUtils")]
