+++ test/TesterInternal/StorageTests/CommonStorageUtilities.cs
