+++ test/TesterSQLUtils/SqlStatisticsPublisherTests/DummyCounter.cs
