+++ test/TesterSQLUtils/SqlStatisticsPublisherTests/MySqlStatisticsPublisherTests.cs
