+++ test/TesterSQLUtils/project.json
@@ -1,0 +1,19 @@
+﻿{
+  "dependencies": {
+    "FluentAssertions": "4.0.0",
+    "xunit": "2.1.0",
+    "xunit.runner.visualstudio": "2.1.0",
+    "Xunit.SkippableFact": "1.2.14",
+    "Microsoft.Data.SQLite": "1.1.0",
+    "Npgsql": "3.1.9",
+    "MySql.Data": "7.0.6-IR31",
+    "system.data.sqlclient" :  "4.3.0"
+  },
+  "frameworks": {
+    "net451": {}
+  },
+  "runtimes": {
+    "win-anycpu": {},
+    "win": {}
+  }
+}
